+++ src/renderer/src/features/home/ui/rival/Rival.tsx
@@ -1,7 +1,7 @@
 import * as S from "./Rival.style";
 import { MyRivalUsers } from "@/features/home/ui/rival/myrival-users/MyRivalUsers";
 import { useRival } from "@/shared/lib/useRival";
-import { AddRivalsDialog, DeleteRivalsDialog } from "@/shared/lib/rival-management";
+import { AddRivalsDialog, DeleteRivalsDialog } from "@/shared/ui/rival-management";
 import { Button } from "@/shared/ui";
 
 export const Rival = () => {
