+++ src/renderer/src/pages/competition/ui/CompetitionPage.tsx
@@ -1,7 +1,7 @@
 import * as S from "./CompetitionPage.style";
 import { useCompetition } from "@/pages/competition/model/useCompetition";
 import { WithMyCompetition, RivalCompetition } from "@/features/competition";
-import { AddRivalsDialog } from "@/shared/lib/rival-management/addRival/AddRivals";
+import { AddRivalsDialog } from "@/shared/ui/rival-management/addRival/AddRivals";
 import { useRival } from "@/shared/lib/useRival";
 import { useState } from "react";
 import { Button, Dialog } from "@/shared/ui";
