--- src/renderer/src/shared/lib/rival-management/index.ts
@@ -1,2 +1,0 @@
-export { AddRivalsDialog } from "@/shared/lib/rival-management/addRival/AddRivals";
-export { DeleteRivalsDialog } from "@/shared/lib/rival-management/deleteRival/DeleteRivals";
