+++ src/renderer/src/shared/ui/rival-management/addRival/AddRivals.style.ts
