+++ src/renderer/src/shared/ui/rival-management/deleteRival/DeleteRivals.style.ts
