+++ src/renderer/src/shared/ui/rival-management/deleteRival/DeleteRivals.tsx
@@ -1,6 +1,6 @@
 import { Button, Dialog } from "@/shared/ui";
 import * as S from "./DeleteRivals.style";
-import { useRival } from "@/shared/lib/useRival";
+import { useRival } from "@/shared/lib";
 
 interface DeleteRivalsDialogProps {
   isOpen: boolean;
