+++ src/renderer/src/shared/ui/rival-management/index.ts
@@ -1,0 +1,2 @@
+export { AddRivalsDialog } from "./addRival/AddRivals";
+export { DeleteRivalsDialog } from "./deleteRival/DeleteRivals";
